requirements.txt
google-generativeai
python-dotenv
feedparser
reportlab
